+++ chap02/src/sec02/exam02/ByteExample.java
