+++ chap02/src/sec02/exam05/StringExample.java
