+++ chap02/src/sec02/exam08/BooleanExample.java
@@ -8,6 +8,5 @@ public class BooleanExample {
 		}else {
 			System.out.println("시작합니다.");
 		}
-
 	}
 }
